rename test package to end in .test
