+++ queryall-lib-test/src/test/java/org/queryall/utils/test/QueryCreatorTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package org.queryall.utils;
+package org.queryall.utils.test;
 
 import java.util.ArrayList;
 import java.util.Collection;
