+++ queryall-lib-test/src/test/java/org/queryall/utils/test/SchemaServiceLoaderTest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package org.queryall.utils;
+package org.queryall.utils.test;
 
 import java.util.HashSet;
 import java.util.Set;
